#60/feature

+++ android/core/data/src/main/java/com/jaino/data/repository/auth/AuthRepository.kt
@@ -9,6 +9,6 @@ interface AuthRepository {
 
     fun setNickName(nickName: String)
 
-    fun setUserId(userId: Int)
+    fun setUserId(userId: Long)
 
 }
